Feature/e2e testing

+++ src/friend-requests/friend-requests.service.ts
@@ -1,5 +1,6 @@
 import {
   ConflictException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
@@ -39,7 +40,7 @@ export class FriendRequestsService {
    *    3. A user can not send a friend request to a user he is already friends with
    *    4. A user can not send a friend request to a user that has sent him a friend request already
    */
-  async sendFriendRequest(sender: User, receiver: User) {
+  async send(sender: User, receiver: User): Promise<FriendRequest> {
     // 1. Rule
     if (sender.id == receiver.id)
       throw new ConflictException('You can not send yourself a friend request');
@@ -94,22 +95,24 @@ export class FriendRequestsService {
       receiver,
     });
 
-    await this.friendRequestRepository.save(friendRequest);
+    return await this.friendRequestRepository.save(friendRequest);
   }
 
-  async acceptFriendRequest(id: string, acceptingUser: User) {
+  async accept(id: string, acceptingUser: User) {
     const friendRequest: FriendRequest = await this.findOne({
       where: {
         id,
-        receiver: {
-          id: acceptingUser.id,
-        },
       },
       relations: {
         sender: true,
         receiver: true,
       },
     });
+
+    if (friendRequest.receiver.id !== acceptingUser.id)
+      throw new ForbiddenException(
+        'You can not accept another users friend request',
+      );
 
     // remove friend request
     await this.friendRequestRepository.remove(friendRequest);
@@ -118,19 +121,21 @@ export class FriendRequestsService {
     await this.usersService.addFriend(acceptingUser, friendRequest.sender);
   }
 
-  async declineFriendRequest(id: string, decliningUser: User) {
+  async decline(id: string, decliningUser: User) {
     const friendRequest: FriendRequest = await this.findOne({
       where: {
         id,
-        receiver: {
-          id: decliningUser.id,
-        },
       },
       relations: {
         receiver: true,
+        sender: true,
       },
     });
 
+    if (friendRequest.receiver.id !== decliningUser.id)
+      throw new ForbiddenException(
+        'You can not decline another users friend request',
+      );
     return await this.friendRequestRepository.remove(friendRequest);
   }
 
@@ -139,20 +144,26 @@ export class FriendRequestsService {
    * @param acceptingUser the user who sent the friend request
    * @param toAcceptUser the user who's friend request should be revoked
    */
-  async revokeFriendRequest(id: string, revokingUser: User) {
+  async revoke(id: string, revokingUser: User): Promise<FriendRequest> {
     const friendRequest: FriendRequest = await this.findOne({
       where: {
         id,
-        sender: {
-          id: revokingUser.id,
-        },
+      },
+      relations: {
+        sender: true,
+        receiver: true,
       },
     });
 
-    await this.friendRequestRepository.remove(friendRequest);
+    if (friendRequest.sender.id !== revokingUser.id)
+      throw new ForbiddenException(
+        'You can not revoke another users friend request',
+      );
+
+    return this.friendRequestRepository.remove(friendRequest);
   }
 
-  async getReceivedFriendRequests(user: User): Promise<FriendRequest[]> {
+  async getReceived(user: User): Promise<FriendRequest[]> {
     return this.friendRequestRepository.find({
       where: {
         receiver: {
@@ -166,7 +177,7 @@ export class FriendRequestsService {
     });
   }
 
-  async getSentFriendRequests(user: User): Promise<FriendRequest[]> {
+  async getSent(user: User): Promise<FriendRequest[]> {
     return this.friendRequestRepository.find({
       where: {
         sender: {
